+++ js/src/auto-instrumentations/bundler/plugin.ts
@@ -33,6 +33,7 @@ import { openRouterConfigs } from "../configs/openrouter";
 import { mistralConfigs } from "../configs/mistral";
 import { cohereConfigs } from "../configs/cohere";
 import { groqConfigs } from "../configs/groq";
+import { gitHubCopilotConfigs } from "../configs/github-copilot";
 
 export interface BundlerPluginOptions {
   /**
@@ -87,6 +88,7 @@ export const unplugin = createUnplugin<BundlerPluginOptions>((options = {}) => {
     ...mistralConfigs,
     ...cohereConfigs,
     ...groqConfigs,
+    ...gitHubCopilotConfigs,
     ...(options.instrumentations || []),
   ];